# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/beers": {
            "get": {
                "description": "Get a list of all beers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Beers"
                ],
                "summary": "Get a list of all beers",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.BeerResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.BeerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new beer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Beers"
                ],
                "summary": "Create a new beer",
                "parameters": [
                    {
                        "description": "Beer input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.BeerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/controllers.BeerResponse"
                        }
                    },
                    "400": {
                        "description": "Ensure input is correct!",
                        "schema": {
                            "$ref": "#/definitions/controllers.BeerErrorResponse"
                        }
                    }
                }
            }
        },
        "/beers/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Beers"
                ],
                "summary": "Get a beer by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Beer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.BeerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.BeerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Beers"
                ],
                "summary": "Delete a beer by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Beer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.DeletedBeerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.BeerErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Beers"
                ],
                "summary": "Update a beer by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Beer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Beer input payload",
                        "name": "beer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.BeerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.BeerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.BeerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.BeerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.BeerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "controllers.BeerInput": {
            "type": "object",
            "required": [
                "beername",
                "brewery"
            ],
            "properties": {
                "beername": {
                    "type": "string"
                },
                "brewery": {
                    "$ref": "#/definitions/controllers.BreweryInput"
                }
            }
        },
        "controllers.BeerResponse": {
            "type": "object",
            "properties": {
                "brewery": {
                    "$ref": "#/definitions/controllers.BreweryInput"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "controllers.BreweryInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "controllers.DeletedBeerResponse": {
            "type": "object",
            "properties": {
                "deleted": {
                    "type": "boolean"
                }
            }
        }
    }
}